map exposed API to where it is imported from
