+++ .vscode/settings.json
@@ -400,6 +400,7 @@
     "toxenv",
     "tpscompare",
     "tqdm",
+    "tred",
     "tsize",
     "tutes",
     "uids",
